Organize modules into appropriate directories
