+++ Causal_Web/gui_pyside/canvas_widget.py
@@ -30,7 +30,7 @@ from ..gui.state import (
     set_selected_connection,
     set_selected_observer,
 )
-from ..command_stack import (
+from ..gui.command_stack import (
     CommandStack,
     MoveNodeCommand,
     MoveMetaNodeCommand,
@@ -534,9 +534,9 @@ class CanvasWidget(QGraphicsView):
                 item.node_id
             )
         elif isinstance(item, EdgeItem):
-            actions[menu.addAction("Delete Connection")] = (
-                lambda: self.delete_connection(item.index, item.connection_type)
-            )
+            actions[
+                menu.addAction("Delete Connection")
+            ] = lambda: self.delete_connection(item.index, item.connection_type)
         elif isinstance(item, ObserverItem):
             actions[menu.addAction("Delete Observer")] = lambda: self.delete_observer(
                 item.index
